requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
Werkzeug==3.0.2
WTForms==3.1.2
